requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
antiorm==1.2.1
anyio==4.3.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
db==0.1.1
db-sqlite3==0.0.1
distro==1.9.0
faiss-cpu==1.8.0
Flask==3.0.2
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.10
langchain-community==0.0.25
langchain-core==0.1.29
langchain-text-splitters==0.0.1
langsmith==0.1.19
MarkupSafe==2.1.5
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.12.0
orjson==3.9.15
packaging==23.2
pluggy==1.4.0
pydantic==2.6.3
pydantic_core==2.16.3
pypdf==4.1.0
PyPDF2==3.0.1
pytest==8.1.0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.28
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
yarl==1.9.4
